requirements.txt
Django~=4.2.5
django-bootstrap5~=23.3
django-htmx~=1.16.0
django-model-utils~=4.3.1
django-money~=3.3.0
django-nested-admin~=4.0.2
django-storages[google]~=1.14.1
environs[django]~=9.5.0
google-cloud-storage~=2.11.0
psycopg2-binary~=2.9.7
Werkzeug~=3.0.0
